+++ Assets/Scripts/Incidents/IncidentActions/IncidentActionHandler.cs
@@ -1,119 +1,70 @@
 ﻿using Sirenix.OdinInspector;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace Game.Incidents
 {
 	public class IncidentActionHandler
 	{
-		public string incidentLog;
-		[ShowInInspector, ListDrawerSettings(CustomAddFunction = "AddNewActionContainer"), HideReferenceObjectPicker]
-		public List<IncidentActionContainer> Actions { get; set; }
-		[ShowInInspector, ListDrawerSettings(CustomAddFunction = "AddNewContextDeployer"), HideReferenceObjectPicker]
-		public List<IContextDeployer> Deployers { get; set; }
-		public Type ContextType { get; set; }
+		[TypeFilter("GetFilteredTypeList"), OnValueChanged("SetAction"), HideLabel]
+		public IIncidentAction incidentAction;
 
-		private IIncidentContext providedContext;
+		private Action onSetCallback;
 
 		public IncidentActionHandler() { }
-		public IncidentActionHandler(Type type)
+		public IncidentActionHandler(Action onSetCallback)
 		{
-			Actions = new List<IncidentActionContainer>();
-			Deployers = new List<IContextDeployer>();
-			ContextType = type;
+			this.onSetCallback = onSetCallback;
 		}
 
-		public bool VerifyActions(IIncidentContext context, Func<int, IIncidentActionField> delayedCalculateAction = null)
+		public IncidentActionHandler(IIncidentAction action, Action onSetCallback) : this(onSetCallback)
 		{
-			providedContext = context;
+			this.incidentAction = action;
+		}
 
-			if (delayedCalculateAction == null)
+		public bool VerifyAction(IIncidentContext context, Func<int, IIncidentActionField> delayedCalculateAction)
+		{
+			if (!incidentAction.VerifyAction(context, delayedCalculateAction))
 			{
-				delayedCalculateAction = GetContextFromActionFields;
+				return false;
 			}
-
-			foreach (var action in Actions)
-			{
-				if (!action.VerifyAction(context, delayedCalculateAction))
-				{
-					OutputLogger.LogWarning("ActionContainer failed to verify action context!");
-					return false;
-				}
-			}
-
 			return true;
 		}
 
-		public bool PerformActions(IIncidentContext context, ref IncidentReport report)
+		public void PerformAction(IIncidentContext context, ref IncidentReport report)
 		{
-			providedContext = context;
-
-			foreach (var action in Actions)
-			{
-				action.PerformAction(context, ref report);
-			}
-
-			report.ReportLog += incidentLog;
-			GetContextDictionary(ref report);
-
-			foreach (var deployer in Deployers)
-			{
-				deployer.Deploy(context, GetContextFromActionFields);
-			}
-
-			return true;
+			incidentAction.PerformAction(context, ref report);
 		}
 
 		public void UpdateActionFieldIDs(ref int startingValue)
 		{
-			if (startingValue == 0)
-			{
-				var constant = new ConstantActionField(ContextType);
-				IncidentEditorWindow.actionFields.Add(constant);
-				startingValue++;
-			}
-
-			foreach(var container in Actions)
-			{
-				container.UpdateActionFieldIDs(ref startingValue);
-			}
+			incidentAction.UpdateActionFieldIDs(ref startingValue);
 		}
 
-		public void GetContextDictionary(ref IncidentReport report)
+		public void AddContext(ref IncidentReport report)
 		{
-			foreach (var action in Actions)
-			{
-				action.AddContext(ref report);
-			}
+			incidentAction.AddContext(ref report);
 		}
 
-		public IIncidentActionField GetContextFromActionFields(int actionFieldID)
+		public bool GetContextField(int id, out IIncidentActionField contextField)
 		{
-			if(actionFieldID == 0)
+			if(incidentAction.GetContextField(id, out contextField))
 			{
-				return new ConstantActionField(providedContext.ContextType);
+				return true;
 			}
 
-			foreach (var action in Actions)
-			{
-				if(action.GetContextField(actionFieldID, out var incidentActionField))
-				{
-					return incidentActionField;
-				}
-			}
-
-			return null;
+			return false;
 		}
 
-		private void AddNewActionContainer()
+		private void SetAction()
 		{
-			Actions.Add(new IncidentActionContainer());
+			incidentAction.UpdateEditor();
+			IncidentEditorWindow.UpdateActionFieldIDs();
 		}
 
-		private void AddNewContextDeployer()
+		private IEnumerable<Type> GetFilteredTypeList()
 		{
-			Deployers.Add(new ContextDeployer());
+			return IncidentActionHelpers.GetFilteredTypeList(IncidentEditorWindow.ContextType);
 		}
 	}
 }
